C++签名工具静态告警清理和代码安全检视 Signed-off-by: liuchang <liuchang281@huawei.com>

+++ hapsigntool_cpp/codesigning/sign/include/code_signing.h
@@ -58,7 +58,7 @@ public:
 
 public:
     static bool IsNativeFile(const std::string& input);
-    uint32_t ComputeDataSize(ZipSigner& zip);
+    int64_t ComputeDataSize(ZipSigner& zip);
     int64_t GetTimestamp();
     bool SignNativeLibs(const std::string &input, std::string &ownerID);
     void UpdateCodeSignBlock();
